+++ config/settings.py
@@ -20,7 +20,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/3.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = "django-insecure-vwbw$25e3ir!b33)=gzjzq#lxxj^0es4e3-t4y@952z3l)$(^("
+SECRET_KEY = "django-insecure-5==!o!#exm0uw)4pb_rkju1!=lbu9j+utag$b)-4!+gq0o%ckg"
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
